Tarone's Z Test

Tests the goodness of fit of the binomial distribution.

Tarone.test(N, M)

Arguments

N

Trials

M

Counts

Value

a htest object

References

https://stats.stackexchange.com/a/410376/6378 and R. E. TARONE, Testing the goodness of fit of the binomial distribution, Biometrika, Volume 66, Issue 3, December 1979, Pages 585–590, https://doi.org/10.1093/biomet/66.3.585

Author

Ben O'Neill

Examples

 #Generate example data
N <- c(30, 32, 40, 28, 29, 35, 30, 34, 31, 39)
M <- c( 9, 10, 22, 15,  8, 19, 16, 19, 15, 10)
Tarone.test(N, M)
#> 
#> 	Tarone's Z test
#> 
#> data:  M successes from N trials
#> z = 2.5988, p-value = 0.009355
#> alternative hypothesis: true dispersion parameter is greater than 0
#> sample estimates:
#> proportion parameter 
#>            0.4359756 
#>