Calculate NOEC Using Many-to-one Pairwise Tests

This function calculates the No Observed Effect Concentration (NOEC) from dose response data using pairwise comparison tests from the rstatix package.

Usage

calculate_noec_rstatix(
  data,
  response,
  dose,
  control = "0",
  test = c("t.test", "wilcox.test"),
  p_adjust_method = "holm",
  alternative = "two.sided",
  alpha = 0.05
)

Arguments

data

A data frame containing the dose response data

response

The name of the response variable (unquoted)

dose

The name of the dose variable (unquoted)

control

The level of the dose variable to be used as control

test

The statistical test to use: "t.test" or "wilcox.test"

p_adjust_method

Method for p-value adjustment for multiple comparisons

alternative

Direction of the alternative hypothesis: "two.sided", "greater", or "less"

alpha

Significance level (default: 0.05)

Value

A list containing the NOEC value and the full test results