get Endpoint for continuous data according to results of a series of tests

get Endpoint for continuous data according to results of a series of tests

Usage

contEndpoint(
  paov,
  pks,
  pnormal,
  phomogeneity,
  monotonicity,
  william,
  dunnett,
  dunn,
  jonckheere,
  procedure = "stepDown",
  doses = c("A", "B", "C", "D")
)

Arguments

paov

Anova p-value

pks

Kolmogorov-Smirnov test p-value

pnormal

Normal test p-value

phomogeneity

Homogeneity variance test p-value

monotonicity

Monotonicity test p-value

william

Williams' test p-value

dunnett

Dunnett's test p-value

dunn

Dunn's test p-value

jonckheere

Jonckheere-Tepstra test p-value

procedure

step down of up

doses

tested doses

Value

NOEC along with an attribute indicating which test was used.

Details

Purpose: The contEndpoint function computes the No Observed Effect Concentration (NOEC) based on the results of various statistical tests. It evaluates these tests' p-values to determine which statistical analysis is most appropriate.The function first checks the significance of the monotonicity test. Depending on the results, it selects the appropriate p-value from the relevant test (Dunnett, Dunn, Williams, or Jonckheere) to compute the NOEC. The function returns the NOEC along with an attribute indicating which test was used. Original function used directly the williams' test or dunnett's test objects as input, which can cause conflicts when using different functions to get the pvals.

Examples

contEndpoint(paov = 0.01, pks = 0.03, pnormal = 0.06,
phomogeneity = 0.07, monotonicity = c(0.1,0.2),
william = c(0.07, 0.06, 0.03, 0.01), dunnett = c(0.07, 0.06, 0.03, 0.01),
dunn = c(0.07, 0.06, 0.03, 0.01),
jonckheere = c(0.07, 0.06, 0.03, 0.01),
procedure = "stepDown",
doses = c("Control","A", "B", "C", "D"))
#> [1] "B"
#> attr(,"test")
#> [1] "Williams"