Calculate Dose Probability for GLMM using PQL

This function calculates the dose probability for a generalized linear mixed model (GLMM) using Penalized Quasi-Likelihood (PQL).

Usage

dose.p.glmmPQL(obj, cf = 1:2, p = 0.5)

Arguments

obj

A fitted GLMM object of class "glmmPQL".

cf

A numeric vector specifying the coefficients to be used for the calculation (default is 1:2).

p

A numeric value representing the probability level (default is 0.5).

Value

A structured object of class "glm.dose" containing the estimated dose probability and its standard error.

Examples

# Assuming `glmm_model` is a fitted glmmPQL model
library(MASS)
glmm_model <- glmmPQL(yt ~ dose,random= ~1 | Obs,family= quasibinomial(link="logit"),data=pvi_example)
#> iteration 1
#> iteration 2
#> iteration 3
#> iteration 4
#> iteration 5
#> iteration 6
#> iteration 7
#> iteration 8
#> iteration 9
#> iteration 10
dose_result <- dose.p.glmmPQL(glmm_model)
print(dose_result)
#>              Dose       SE
#> p = 0.5: 21.22061 1.455262