added functionality for mselect

added functionality for mselect

Deprecated helper function for mselect

Usage

mselect.plus(
  object = NULL,
  fctList = NULL,
  nested = FALSE,
  sorted = c("IC", "Res var", "Lack of fit", "no"),
  linreg = FALSE,
  icfct = AIC,
  respCol = "effect",
  doseCol = "dose",
  data = NULL,
  type = "continuous",
  additionalReliability = c("EFSA")
)

mselect.ZG(
  object = NULL,
  fctList = NULL,
  nested = FALSE,
  sorted = c("IC", "Res var", "Lack of fit", "no"),
  linreg = FALSE,
  icfct = AIC,
  respCol = "effect",
  doseCol = "dose",
  data = NULL,
  type = "continuous",
  additionalReliability = c("EFSA")
)

Arguments

object

a fitted object of class 'drc'.

fctList

a list of dose-response functions to be compared.

nested

logical. TRUE results in F tests between adjacent models (in 'fctList'). Only sensible for nested models.

sorted

character string determining according to which criterion the model fits are ranked, default is IC.

linreg

logical indicating whether or not additionally polynomial regression models (linear, quadratic, and cubic models) should be fitted (they could be useful for a kind of informal lack-of-test consideration for the models specified, capturing unexpected departures).

icfct

function for supplying the information criterion to be used. AIC and BIC are two options.

respCol

name of the response column

doseCol

name of the dose column

data

data used

type

type of models, binomial, continuous, etc.

additionalReliability

additional reliability need to be calculated

Value

a model comparison object with class drcComp

Details

Due to old ECxHelper development where mselect.plus is defined as mselect.ZG

Examples

if (FALSE) { # \dontrun{
data("dat_medium")
dat_medium <- dat_medium %>% mutate(Treatment=factor(Dose,levels=unique(Dose)))
dat_medium$Response[dat_medium$Response < 0] <- 0
mod <- drm(Response~Dose,data=dat_medium,fct=LL.3())
fctList <- list(LN.4(),LL.4(),W1.3(),LL2.2())
res <- mselect.plus(mod,fctList = fctList )
modList <- res$modList
} # }