Plot a list of models together.

Plot a list of models together.

Usage

# S3 method for class 'modList'
plot(
  modList,
  respLev = NULL,
  data = NULL,
  xmin,
  xmax,
  scale = c("logx", "logy", "logxy", "orig"),
  npts = 100,
  plot_respLev = FALSE,
  xbreaks = NULL,
  ymin = NULL,
  ymax = NULL,
  ...
)

Arguments

modList

list of drc models.

respLev

to be calculated ECx levels, for example, c(10,20,50).

data

data used to fit the set of models

xmin

minimum value of xaxis, should be greater than 0 to fit the log scale.

xmax

maximum value of xaxis, can be missing.

scale

could be "logx", "logy", "logxy", "orig".

npts

number of points used for plotting the prediction confidence bands

plot_respLev

logical, whether to add the estimated ECx with CIs

xbreaks

breaks in x-axis labeling

...

addition parameters to be passed into plotting functions or user input when the use does not want to use the data provided in the modList

Value

a ggplot object

Examples

if (FALSE) { # \dontrun{
} # }