Preliminary Plot 1 for Dose Response Data

This function generates a scatter plot of response against nominal dose.

Usage

prelimPlot1(
  testdata,
  dose_col = "Dose",
  response_col = "Response",
  ylab = "Response",
  xlab = "Test Concentration [nominal, mg a.s./L]",
  title = "Measured Variable"
)

Arguments

testdata

A data frame containing the dose and response data.

dose_col

name of the dose column, default being "Dose".

response_col

name of the response column.

ylab

A string for the y-axis label. Default is "Response".

xlab

A string for the x-axis label. Default is "Test Concentration [nominal, mg a.s./L]".

title

A string for the plot title. Default is "Measured Variable".

Value

A ggplot object.