Function to test for overdispersion in binomial data using glm approach

Function to test for overdispersion in binomial data using glm approach

Usage

test_overdispersion(successes, totals, model = NULL)

Arguments

successes

Numeric vector of successes (e.g., number of survivors).

totals

Numeric vector of total observations (e.g., total organisms).

model

A fitted binomial GLM with predictors accounted for. (default is NULL)

Value

a list of dispersion test results