Obtain Endpoint (NOEC) according to a series of p-values

Obtain Endpoint (NOEC) according to a series of p-values

Usage

getEndpoint(pvals, doses = c("Control", "B", "C", "D"), procedure = "stepDown")

Arguments

pvals

pvals from a tests

doses

corresponding doses

procedure

procedure to obtain NOEC

Value

NOEC

Details

If all p-values are NA, it initializes them to zeros. If there are fewer p-values than doses, it pads the p-values with zeros.It checks which doses are significant by comparing the p-values to a threshold (0.05).Depending on the specified procedure, it either steps down from the highest dose to find the NOEC or steps up from the lowest dose..

Examples

pvals <- c(0.01, 0.03, 0.07)
doses <- c("Control", "B", "C", "D")
getEndpoint(pvals, doses)
#> [1] "D"