Plot a list of models together.

Plot a list of models together.

Usage

# S3 method for class 'modList'
plot(
  x,
  respLev = NULL,
  data = NULL,
  xmin,
  xmax,
  scale = c("logx", "logy", "logxy", "orig"),
  npts = 100,
  plot_respLev = FALSE,
  xbreaks = NULL,
  ymin = NULL,
  ymax = NULL,
  ...
)

Arguments

x

a list of drc models with modList class.

respLev

to be calculated ECx levels, for example, c(10,20,50).

data

data used to fit the set of models

xmin

minimum value of xaxis, should be greater than 0 to fit the log scale.

xmax

maximum value of xaxis, can be missing.

scale

could be "logx", "logy", "logxy", "orig".

npts

number of points used for plotting the prediction confidence bands

plot_respLev

logical, whether to add the estimated ECx with CIs

xbreaks

breaks in x-axis labeling

ymin

minimum of y

ymax

max of y

...

addition parameters to be passed into plotting functions or user input when the use does not want to use the data provided in the modList

Value

a ggplot object

Examples

if (FALSE) { # \dontrun{
data("dat_medium")
dat_medium <- dat_medium %>% mutate(Treatment=factor(Dose,levels=unique(Dose)))
dat_medium$Response[dat_medium$Response < 0] <- 0
mod <- drm(Response~Dose,data=dat_medium,fct=LL.3())
fctList <- list(LN.4(),LL.4(),W1.3(),LL2.2())
res <- mselect.plus(mod,fctList = fctList )
modList <- res$modList
p <- plot.modList(modList[1:3])
p
} # }